+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## [1.0.0-rc.20](https://github.com/Arekkazu/SGPMP-FRONT-END-PWA/compare/v1.0.0-rc.19...v1.0.0-rc.20) (2026-09-08)
+
+### Features
+
+* **rf25:** cargar roles dinámicos en el modal de usuario en vez de quemados ([16d4fec](https://github.com/Arekkazu/SGPMP-FRONT-END-PWA/commit/16d4fec05a90aaef10192352cf6b2fe65df25298))
+* **rf25:** selección de fincas asignadas en el detalle de usuario (RF-25) ([3295a89](https://github.com/Arekkazu/SGPMP-FRONT-END-PWA/commit/3295a896a72a6d7519483e6defa63eab92b22f76))
+
+### Bug Fixes
+
+* **auth:** ocultar el login SSO de AgroFusion ([b0bbc81](https://github.com/Arekkazu/SGPMP-FRONT-END-PWA/commit/b0bbc819782e390a3608c3e831bfc7535bceec47))
+
+## [1.0.0-rc.19](https://github.com/Arekkazu/SGPMP-FRONT-END-PWA/compare/v1.0.0-rc.18...v1.0.0-rc.19) (2026-09-08)
+
+### Bug Fixes
+
+* **rf16:** consumir catalogos de especie en activos ([951b76b](https://github.com/Arekkazu/SGPMP-FRONT-END-PWA/commit/951b76bcd0789763a1b39d0b877ecac552cd0733))
+
 ## [1.0.0-rc.18](https://github.com/Arekkazu/SGPMP-FRONT-END-PWA/compare/v1.0.0-rc.17...v1.0.0-rc.18) (2026-09-08)
 
 ### Features